PRUEBA 1

+++ oop2_sportConnect/src/main/java/co/edu/ue/repository/IEventoJpa.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import co.edu.ue.model.Eventos;  
 
-public interface IEventoJpa extends JpaRepository<Eventos, Long> {  // Cambiado el ID a Long
+public interface IEventoJpa extends JpaRepository<Eventos, Integer > {  
 }
